Move the CoreML EP selection tests to test_selection_coreml.cc
Move the CoreML EP selection tests from test_selection.cc to a dedicated CoreML test file.
